+++ Sources/VCore/Helpers/Architectural Pattern Helpers/SwiftUI/Alert/AlertButtonConvertible.swift
@@ -9,7 +9,6 @@ import SwiftUI
 
 // MARK: - Alert Button Convertible
 /// Type that allows for conversion to `AlertButtonProtocol`.
-@MainActor
 public protocol AlertButtonConvertible {
     /// Converts self to `AlertButtonProtocol` `Array`.
     func toButtons() -> [any AlertButtonProtocol]
@@ -20,5 +19,5 @@ extension Array: AlertButtonConvertible where Element == any AlertButtonProtocol
 }
 
 extension EmptyView: AlertButtonConvertible {
-    public func toButtons() -> [any AlertButtonProtocol] { [] }
+    nonisolated public func toButtons() -> [any AlertButtonProtocol] { [] }
 }