+++ Sources/VCore/Helpers/Architectural Pattern Helpers/SwiftUI/Confirmation Dialog/ConfirmationDialogButtonConvertible.swift
@@ -9,7 +9,6 @@ import SwiftUI
 
 // MARK: - Confirmation Dialog Button Convertible
 /// Type that allows for conversion to `ConfirmationDialogButtonProtocol`.
-@MainActor
 public protocol ConfirmationDialogButtonConvertible {
     /// Converts self to `ConfirmationDialogButtonProtocol` `Array`.
     func toButtons() -> [any ConfirmationDialogButtonProtocol]
@@ -20,5 +19,5 @@ extension Array: ConfirmationDialogButtonConvertible where Element == any Confir
 }
 
 extension EmptyView: ConfirmationDialogButtonConvertible {
-    public func toButtons() -> [any ConfirmationDialogButtonProtocol] { [] }
+    nonisolated public func toButtons() -> [any ConfirmationDialogButtonProtocol] { [] }
 }