+++ Sources/VCore/Helpers/Architectural Pattern Helpers/SwiftUI/Confirmation Dialog/ConfirmationDialogButtonProtocol.swift
@@ -14,6 +14,7 @@ public protocol ConfirmationDialogButtonProtocol: ConfirmationDialogButtonConver
     typealias Body = AnyView
 
     /// Creates a `View` that represents the body of a button.
+    @MainActor
     func makeBody(
         animateOutHandler: @escaping (/*completion*/ (() -> Void)?) -> Void
     ) -> Body