+++ Sources/VCore/Helpers/Architectural Pattern Helpers/UIKit/UI Action Sheet/UIActionSheetButtonBuilder.swift
@@ -18,42 +18,34 @@ public struct UIActionSheetButtonBuilder: Sendable {
     public typealias Result = [any UIActionSheetButtonProtocol]
     
     // MARK: Build Blocks
-    @MainActor
     public static func buildBlock() -> Result {
         []
     }
     
-    @MainActor
     public static func buildBlock(_ components: Component...) -> Result {
         components.flatMap { $0.toButtons() }
     }
     
-    @MainActor
     public static func buildOptional(_ component: Component?) -> Result {
         component?.toButtons() ?? []
     }
     
-    @MainActor
     public static func buildEither(first component: Component) -> Result {
         component.toButtons()
     }
     
-    @MainActor
     public static func buildEither(second component: Component) -> Result {
         component.toButtons()
     }
     
-    @MainActor
     public static func buildArray(_ components: [Component]) -> Result {
         components.flatMap { $0.toButtons() }
     }
     
-    @MainActor
     public static func buildLimitedAvailability(_ component: Component) -> Result {
         component.toButtons()
     }
     
-    @MainActor
     public static func buildFinalResult(_ component: Component) -> Result {
         component.toButtons()
     }