+++ Sources/VCore/Helpers/Architectural Pattern Helpers/UIKit/UI Action Sheet/UIActionSheetButtonConvertible.swift
@@ -11,7 +11,6 @@ import UIKit
 
 // MARK: - UI Action Sheet Button Convertible
 /// Type that allows for conversion to `UIActionSheetButtonProtocol`.
-@MainActor
 public protocol UIActionSheetButtonConvertible {
     /// Converts self to `UIActionSheetButtonProtocol` `Array`.
     func toButtons() -> [any UIActionSheetButtonProtocol]