+++ Sources/VCore/Helpers/Architectural Pattern Helpers/UIKit/UI Action Sheet/UIActionSheetButtonProtocol.swift
@@ -13,6 +13,7 @@ import UIKit
 /// `UIActionSheetController` button protocol.
 public protocol UIActionSheetButtonProtocol: UIActionSheetButtonConvertible {
     /// Converts `UIActionSheetButtonProtocol` to `UIAlertAction`.
+    @MainActor
     func makeBody() -> UIAlertAction
 }