+++ Sources/VCore/Services and Managers/Keyboard Observer/KeyboardObserver.swift
@@ -88,7 +88,6 @@ public final class KeyboardObserver: ObservableObject { // TODO: iOS 17.0 - Conv
     // MARK: Keyboard Management
 #if canImport(UIKit) && !os(watchOS)
     
-    @MainActor
     private func keyboardWillShow(notification: Notification) {
         let systemKeyboardInfo: SystemKeyboardInfo = .init(notification: notification)
         
@@ -137,7 +136,6 @@ public final class KeyboardObserver: ObservableObject { // TODO: iOS 17.0 - Conv
         }
     }
 
-    @MainActor
     private func keyboardWillHide(notification: Notification) {
         let systemKeyboardInfo: SystemKeyboardInfo = .init(notification: notification)